requirements.txt
wait-for-it==2.2.0
pytest==6.2.5
pre-commit==2.15.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
typing-extensions==4.2.0
pydantic==1.10.7
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.63.0
orjson==3.8.10
uvicorn==0.12.2
httptools==0.5.0
gunicorn==20.1.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
backoff==2.2.1
pyjwt==2.7.0
httpx==0.24.1